random service time in queue, updated sensors

+++ QueueManagementSystem/QueueManagementSystem/src/Model/Sensors/NumericSensor.cs
@@ -1,6 +1,6 @@
 ﻿
 namespace QueueManagementSystem.Model.Sensors {
 	interface NumericSensor {
-		float MeasurePerson(Person person);
+		double MeasurePerson(Person person);
 	}
 }
